chore(design): AgentChatInput.Root owns the composer's bridge and draft state (#8674)
